backend/requirements.txt
Django==5.2.9
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.3.0
python-dotenv==1.0.1